requirements.txt
# Command o install - pip install -r requirements.txt
streamlit
python-dotenv #meant for local variables
langchain
langchain-google-genai
google-generativeai
pypdf